[phase2/weave] Use 4 separate environment files
This allows for better separation, i.e. `base.env` should be static,
and `images.env` contains only image variables and can be overwritten
from a single template with Ansible or other tools. And `local.env`
is there for overriding all the things in any fashion and anything
that would be host-specific and doesn't have to set (if there are
such things these days anyway).