Use HTTPS to download executables in provision.sh
`git.io` supports HTTPS, and `provision.sh` should use it. Otherwise, it would be trivial for an attacker to launch a man-in-the-middle (MITM) attack against the machine running the script. Since the files being downloaded are executed by the same script promptly thereafter, this is a very easy way for an attacker to achieve remote code execution on the machine running the script.