Fix race condition when waiting for kubeconfig.
If WAIT_FOR_KUBECONFIG is set and the cluster comes up very quickly so
that first ssh call succeeds, then the output from ssh's initialization
(generating keys, showing fingerprints/randomart, etc.) can get included
at the start of the kubeconfig.json file, rendering it invalid.

To alleviate this, echo a marker to indicate the start of the file so we
can delete this noise.