In GCE, make sure apt-transport-https is installed.
I found that this step was required when testing with a Debian image
(debian-8-jessie-v20170426), and it's also one the steps in the kubeadm
getting started guide.

This doesn't actually allow the Debian image to be used end-to-end,
since there are still dependency conflicts with the Docker package, but
it does allow the startup to get a little further.