phase2: convert KUBEADM_KUBERNETES_VERSION to semantic version
The script has to know the MINOR version from a k8s semantic version,
so that it can decide which kubeadm config version to use.
Fetch the semantic version from the server if it's not in semantic
format already.
The raw $KUBEADM_KUBERNETES_VERSION can be passed to the config
as kubeadm can handle that.